chore: 상수 constants화 및 동적으로 값 관리
