+++ src/components/Common/PartRecruit.tsx
@@ -1,29 +1,14 @@
-import PmDesign from "@/../public/assets/images/pm-desing-image.svg";
-import Front from "@/../public/assets/images/front-image.svg";
-import Back from "@/../public/assets/images/back-image.svg";
-import Image from "next/image";
+import Image, { StaticImageData } from "next/image";
 
-export default function PartRecruit() {
-  const contents = [
-    {
-      title: "기획 및 디자인",
-      content:
-        "프로젝트의 디자인을 담당하는 파트이며, 디자인 툴을 익히고 디자인 이론과 용어들로 협업에 도움이 되는 스킬을 키울 수 있어요.",
-      image: PmDesign,
-    },
-    {
-      title: "프론트엔드",
-      content:
-        "사용자와 가까이 있는 개발 파트로, 개발 언어를 학습하고 개발 역량을 키웁니다. 개발을 위한기초부터 심화까지 배워요.",
-      image: Front,
-    },
-    {
-      title: "백엔드",
-      content:
-        "눈에 보이지 않는 서버 담당하는 백엔드는 서비스의 요구사항에 맞춰 API를 개발하고 배포 및 운영을 통해 데이터를 관리해요.",
-      image: Back,
-    },
-  ];
+interface PartRecruitProps {
+  contents: {
+    title: string;
+    content: string;
+    image: StaticImageData;
+  }[];
+}
+
+export default function PartRecruit({ contents }: PartRecruitProps) {
   return (
     <div className="flex flex-col mt-[268px]">
       <strong className="text-white text-[24px] font-bold text-center">
