+++ src/components/Recruit/FAQ.tsx
@@ -2,34 +2,14 @@
 
 import { useState } from "react";
 
-export default function FAQ() {
-  const questions = [
-    {
-      question: "공식적인 정기 모임은 언제인가요?",
-      answer:
-        "멋쟁이 사자처럼 ERICA 13기 1학기는 매주 목요일 오후 7시 - 9시에 정기 세션을 진행합니다. 정기 세션은 매주 필수로 참여해야 수료가 가능하며, 매주 참여가 어려울 경우 최종 선발이 어려울 수 있습니다.",
-    },
-    {
-      question: "회비는 얼마인가요?",
-      answer:
-        "멋쟁이 사자처럼 ERICA 13기 회비는 5만원입니다! (4만 원 + 보증금 1만 원)",
-    },
-    {
-      question: "2학기에도 아기사자를 모집하나요?",
-      answer:
-        "2학기에는 아기사자를 모집하지 않습니다. 1년에 한 번, 2월 ~ 3월 모집 기간에만 아기사자를 모집하고 1년 동안 동아리 활동이 진행됩니다.",
-    },
-    {
-      question: "휴학생이어도 지원할 수 있나요?",
-      answer: "재학생, 휴학생, 졸업 유예이시면은 지원 가능합니다.",
-    },
-    {
-      question: "경험과 실력이 부족한데도 들어가도 괜찮을까요?",
-      answer:
-        "물론입니다! 학년, 전공 상관없이 본인의 아이디어를 IT 서비스로 실현시키고자 하는 열정 가득한 분이라면 누구나 지원 가능합니다. 기초부터 차근차근 교육을 진행하니 너무 걱정하지 않으셔도 괜찮습니다.",
-    },
-  ];
+interface FAQProps {
+  questions: {
+    question: string;
+    answer: string;
+  }[];
+}
 
+export default function FAQ({ questions }: FAQProps) {
   const [rotations, setRotations] = useState(
     Array(questions.length).fill(false)
   );
