+++ src/constants/schedule-recruit.ts
@@ -1,0 +1,24 @@
+const roads = [
+  {
+    schedule: "2.7~2.21",
+    title: "서류 모집",
+  },
+  {
+    schedule: "2.23",
+    title: "서류 합격자 발표",
+  },
+  {
+    schedule: "2.24~2.28",
+    title: "면접 진행",
+  },
+  {
+    schedule: "3.2",
+    title: "최종 합격자 발표",
+  },
+  {
+    schedule: "3.4",
+    title: "OT",
+  },
+];
+
+export { roads };
